IncompleteTimestamp: typedef around Poco::Timestamp
The gateway application can encounter a situation when after boot
there is no actual date-time available. The reasons are:

* no battery for RTC module
* no NTP sync available (e.g. network is down)

This does not stop the sensor data collection to work, however,
an incomplete time is sampled. The incomplete time can be fixed
when the time is set up e.g. via the NTP later. Until the system
time is fixed, collect incomplete timestamps (time since 1. 1. 1970).

We can fix a timestamp by treating the incomplete timestamp as an
offset starting at 0. If we know the uptime and current time then
we can fix the timestamp by simple arithmetics.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>