CommandHandler: new class
The CommandHandler interface implements procedures to fullfill
requested Commands. Each Command holds information about some action to
be performed. The action itself can be realized in many different ways.
E.g. command that asks to unpair a device is implemented differently
for each sensoric network.

Classes that want to react on commands must implement this interface
and register with a particular instance of CommandDispatcher class
(introduced in the following commits).

Signed-off-by: Petr Kohout <xkohou14@stud.fit.vutbr.cz>
Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>