AnswerQueue: new class
When executing multiple commands we have to asynchronously wait for
their results. All results of one command are hold by an instance of
the Answer class. Thus, for multiple commands there are multiple Answer
instances that must be watched for results.

The AnswerQueue is a common mechanism that allows to wait for single or
multiple Answer instances easily just by calling queue.wait(). All
Answers with (partial or final) results are scanned this way.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>