CommandDispatcher: new class
One of the core parts of the Gateway is handling of commands coming
from users. The CommandDispatcher contains a set of CommandHandler
instances that can handle certain Command types. It dispatches commands
to the handlers.

There is usually one CommandDispatcher in the system as a central point
for sending commands. Executing of commands is asynchronous.

Signed-off-by: Petr Kohout <xkohou14@stud.fit.vutbr.cz>
Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>