ZWaveDriver: new class
The Driver class handles communication between OpenZWave and
a device attached via a serial port (typically a controller).
Class requires the name of the USB device to be opened.
If the requested driver cannot be loaded program is terminated.

After the call of unregisterItself() method OpenZWave library
prints statistics from registered driver.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>