tools: generate Gateway certificate and key-pair
The gateway application requires to have a valid private key for self
identification and a certificate proving that it has the ID it proclaims
to have. This mechanism is used for deploying gateways however it makes
testing complicated.

To simplify the situation, this script generates the appropriate key-pair
and certificate. There are two usual use cases:

1. I want to quickly test the gateway in a testing environment and I have
   no requirements on the certificate. Solution (I need to have some valid
   gateway ID):

   $ tools/gateway-gencert.sh -G <ID>
   $ ls
   gateway.pem  gateway.pair ...

2. I have set up the testing environment (with a server) and I need a
   certificate that is acceptable by the testing server. The server
   trusts to a certification authority represented by a given key-pair.
   Solution:

   $ tools/gateway-gencert.sh -G <ID> -c ca.key -k gateway.pubkey
   $ ls
   gateway.pem  gateway.pair  gateway.pubkey ...

   The public key is likely unnecessary and can be used to verify signitures
   coming from such gateway or to encrypt data for the specific gateway.

The gateway ID must be generated by some external tool at the moment.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>