AbstractDistributor: data collecting capability implementation
Integration of AbstractDistributor with DistributorListener
for data collection.

Implemets listener registration and notification.
Notification is via "notifyListeners()" protected method. This method is
meant to be called at the beginning of "export()" method in specific Distributor
implementation.

This mechanism is mainly used for collecting data, their analysis
and statistics computation.

As this is not the primary objective of Distributor, notification of
listeners happens separately via AsyncExecutor, so data could be shipped to
Exporters as soon as posible.

BasicDistributor and default configuration files were also updated.

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>