TestingCenter: new class
TestingCenter is a facility intended for injecting commands into
the Gateway runtime without an active remote server. The class
is intended for testing ONLY.

The current configuration initializes a TCP console at localhost:6000.
The testing center is disabled as default. To enable, start the gateway
as

 $ beeeon-gateway -c conf/gateway-startup.ini -Dtesting.center.enable=yes

The TestingCenter is compiled into a separate library. Thus, it can
be easily omitted from installation.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>