CommandSender: new class
CommandSender introduces a common set of methods for implementing a
class that dispatches commands into the system. However, an instance
that sends commands can at the same time implement the CommandHandler
interface. This can lead into a situation when the CommandDispatcher
dispatches a command back to its sender.

Sending commands to itself is considered as an unwanted behaviour.
The CommandSender allows to prevent it safely in its dispatch method.
Thus, all classes that needs to dispatch commands should inherit from
CommandSender.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>