PocoAnswerImpl: extract poco-specific details
To decouple poco-specific logic from the Answer class, the
PocoAnswerImpl is provided here. Later, the PocoAnswerImpl is
generalized even more to make the Answer independent on the backend
implementation.

Tasks are now managed by AutoPtr. The PocoAnswerImpl is an owner of all
its tasks. We can avoid clearing the task list - it is now valid all the
time until the PocoAnswerImpl is destroyed.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>