JablotronDeviceManager: implement listen
Implement the GatewayListenCommand. New devices are detected in the
main loop, however, ignored when not in the listen mode. The listen
mode is created temporarily and finished by the Poco::Timer calling the
stopListen() callback.

All paired devices are stored in map, where DeviceID is key. The second
item is SharedPtr to JablotronDevice, that represents management of
given device. If listen mode is initiated and some message arrives
from devices that do not have created JablotronDevice, JablotronDevice
is created and this device is marked as unpaired. Listen mode is
finished and then, nullptr is set to all devices that are not paired,

Only devices associated with the Jablotron dongle can be discovered and
paired. Association of new unknown devices is currently not possible
because the BeeeOn system is currently missing a mechanism for this
purpose.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>