JablotronDeviceManager: implement unpair
Unpaired devices are set to nullptr. This denotes that such device
is associated with the dongle but unpaired from the BeeeOn system.

An unpaired device is not deassociated from the dongle because the
current BeeeOn system does not allow to pair an unassociated device.
The association requires to know the device's serial number in advance.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>