CommandDispatcherTest: fix elapsed time
After executing of the command, operations that lasted for a certain
period of time were performed. This delay led to the incorrect
verification, that thread was run in the required time.

This is just a quick fix to avoid failing. The test should be
simplified in the future to avoid any timing issues to ever happen.

Fixes: 5780b5dd("CommandDispatcher: new class")

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>