DongleDeviceManager: new abstract class
Provide a generic mechanism for detecting of a dongle. Each device manager
that communicates via some dongle that can be hot-plugged should implement
this class. It defines 2 execution states:

* dongleMissing - executed when no appropriate dongle is available
* dongleAvailable - executed when the appropriate dongle is connected

The device manager implementation thus does not have to care about the
dongle availability. The only case when it should care is when some
communication fails. Any exception thrown from the dongleAvailable()
routines leads to verification whether the appropriate dongle is there.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>