DeviceManager: implement deviceList
Implements sending of ServerDeviceListCommand with waiting for the
response. Almost every DeviceManager implementation performs this
command on startup and it is default non-blocking. This implementation
should provide this operation for general use.

At least one of the Results from Answer must be in SUCCESS state to
obtain device list, throws Exception otherwise.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>