DefaultZWaveDeviceInfo: new class
The class represents common converting of data from OpenZWave
library to BeeeOn system and vice versa.

The class is used in case that device does not have registered class
for converting.

Current implementation uses for converting of data from OpenZWave
library to BeeeOn system only conversion to number. Value that
will be set is converted to string and it is sent from
BeeeOn system to OpenZWave library.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>