DongleDeviceManagerTest: fix race-condition
In some cases, the DongleDeviceManagerTest might fail without any
objective reasons. This leads to false negatives while testing:

  ---
  file: 'test/core/DongleDeviceManagerTest.cpp'
  line: 130
  message: 'assertion failed'
  detail: 'Expression: m_manager.done().tryWait(10000)'
  ---
not ok 9 - N6BeeeOn23DongleDeviceManagerTestE::testNoDongleRun

The situation arises when leaving the dongleMissing() method.
Then there is a short time-window when it is possible to toggle
the m_stop variable. The test however does not count with this
situation and assumes that the DongleDeviceManager::run() loop
would again enter the dongleMissing() routine and sleep there.

Thus the Event becameMissing() has been added to make sure that
the DongleDeviceManager has entered the dongleMissing(). This
should avoid the reported race condition.

Waiting for becameMissing() event has been added to all places
where we expect entering the dongleMissing() routine to avoid
any other related race-conditions.

Fixes: 0d3b919e62e2 ("DongleDeviceManager: new abstract class")

Reported-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
[tested by running 5000 times in cycle]
Tested-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>