MosquittoExporter: new exporter
Exporter connects during the first call to ship().

When the mosquitto library returns some non-zero code the throw is
called and the MosquittoExporter is disconnected. In next calling
to ship() it tries to connect again.

Some MQTT broker is required for the MosquittoExporter to run. This
exporter is only joining an already running broker.

Be aware that it is not possible to have the same client name for more
exporters connected to the same broker. The provided configuration
would always work with only a single MosquittoExporter.

This patch requires libraries libmosquitto and libmosquittopp.

Signed-off-by: Martin Sakin <xsakin00@stud.fit.vutbr.cz>
[fixed commit log]
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>