JablotronDeviceManager: integrate JablotronDeviceAC88
The last measured values of AC-88 device are obtained
from JablotronDeviceManager. Only this type of device
enables to set a value.

Two AC-88 devices can be paired to one USB dongle.
Their setting is performed using common message.
The status of both devices should not be changed
in one message. First of all, status of the first
device is changed and then status is changed on the second device.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>