VirtualModule: new class
Virtual module is part of virtual device. It is defined by module
identifier. It has type and attributes alternatively. It is able to
generate values in various way (constant, random and sinus generator are
supported) and in various range. Value can be also set to virtual module.

Types of module
* Module is called sensor if generator is defined and reaction is set
to none.
* Module is called actuator if generator is not enabled and reaction
is set to success or fail.
* Combination of mentioned setting is allowed but it is not supported
within the system.

Signed-off-by: Klara Necasova <xnecas24@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>