VirtualDeviceManager: new class
It ensures loading of virtual devices from configuration file.
Virtual device that is loaded has to have "enabled" set to "yes"
in configuration file. It is able to send NewDeviceCommand to
CommandDispatcher. It can also send values from modules of
registered devices. These devices are stored in queue and
calendar serves for planning of value sending.
The virtual device manager is able to react to ListenCommand,
DeviceAcceptCommand, SetValueCommand and UnpairCommand.

Signed-off-by: Klara Necasova <xnecas24@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>