VirtualDeviceManager: integrate into gateway
This commit introduces 3 testing virtual devices:

1. It simulates a temperature and humidity sensor. It generates
constant temperature 5 degrees Celsius and random humidity
in range from 0 to 15 percent.

2. This virtual device is not enabled, it does not generate values.
On condition that "enable" was set to "yes" in configuration file,
virtual device simulates a pressure sensor. It generates values
from 0 to 15 Pascals using sinus generator.

3. This virtual device is not paired, it does not generate values.
On condition that "paired" was set to "yes" in configuration file
or device is paired, virtual device simulates a motion sensor.
It generates values 0 or 1 randomly.

4. It simulates a plug (an actuator). Setting of value is always successful.

5. It simulates a plug (an actuator). Setting of value is always unsuccessful.

Signed-off-by: Klara Necasova <xnecas24@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>