CommandHandler: drop name memeber
The name() is useless. There was the only use of this member in
the CommandDispatcher where it reports a duplicate handler
registration. However, the name always informs only about a kind
of class name. But we would rather report the name of the instance.
In fact, this is so minor issue because the number of handlers will
always be very small. So, there is no reason to report any details
about the duplicate problem.

Moreover, It is no more then the name of the inherited class. We
can get this name via ClassInfo if we want. So, there is really no
reason to preserve it anymore.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>