config: define alias hotplugMonitor
The hotplug events an be delivered by two implementations:
UDevMonitor and PipeHotplugMonitor. They can coexist however
to save resources and to allow to disable udev, it is currently
better to select only one of those implementations.

The selection can be done via configuration option hotplug.impl.
The default hotplug monitor is UDevMonitor. To switch to the
PipeHotplugMonitor, start the Gateway as:

  $ beeeon-gateway -c conf/gateway-startup.ini ... \
	  -Dhotplug.impl=pipeHotplug

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>