CommandSender: break circular dependency on CommandDispatcher
Very often it happens that a class derivied from CommandSender also
plays role of CommandHandler and thus the same CommandDispatcher
is either registered into the CommandSender and the sender itself
is registered into the CommandDispatcher. This introduces a circular
dependency that leads to undeletable memory because of using SharedPtr.

We solve this by using UnsafePtr that provides not-NULL checks but
does not hold any reference and nor deletes any memory.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>