DongleDeviceManager: detect failing dongle
When the dongleAvailable() method throws an exception and the dongle
is still connected, we are quickly calling the dongleAvailable() method
again. This however does not improve the sitation as we are now looping
fast inside the DongleDeviceManager without any possibility to recover.

Instead, utilize the FailDetector to detect a number of fails following
quickly. If the dongle fails multiple times quickly, then prevent calling
dongleAvailable() for a while. Instead, the method dongleFailed is called.
Its purpose is to delay the next attempt to call dongleAvailable() or
to wait until the dongle is removed from the device.

The patch has been tested manually on a PC without hciconfig where the
BluetoothAvailabilityManager fails to initialize the HCI interface.
Without this patch, we get a very high rate of failures many times
per second:

15:35:39.191 21281 [Debug] registering dongle bluetooth,,host,hci0,
15:35:39.191 21281 [Information] dongle is available: hci0
15:35:39.193 21281 [Error] I/O error: hciconfig hci0 up has failed: 72
15:35:39.193 21281 [Information] dongle is available: hci0
15:35:39.194 21281 [Error] I/O error: hciconfig hci0 up has failed: 72
15:35:39.194 21281 [Information] dongle is available: hci0
15:35:39.195 21281 [Error] I/O error: hciconfig hci0 up has failed: 72
15:35:39.195 21281 [Information] dongle is available: hci0
15:35:39.196 21281 [Error] I/O error: hciconfig hci0 up has failed: 72
15:35:39.196 21281 [Information] dongle is available: hci0
15:35:39.197 21281 [Error] I/O error: hciconfig hci0 up has failed: 72
15:35:39.197 21281 [Information] dongle is available: hci0
15:35:39.198 21281 [Error] I/O error: hciconfig hci0 up has failed: 72
15:35:39.198 21281 [Information] dongle is available: hci0
15:35:39.199 21281 [Error] I/O error: hciconfig hci0 up has failed: 72
15:35:39.199 21281 [Information] dongle is available: hci0
15:35:39.200 21281 [Error] I/O error: hciconfig hci0 up has failed: 72

After applying this patch, the errors are muted after the given amount
of retries (default is 3).

15:33:33.118 20071 [Debug] registering dongle bluetooth,,host,hci0,
15:33:33.118 20071 [Information] dongle is available: hci0
15:33:33.120 20071 [Error] I/O error: hciconfig hci0 up has failed: 72
15:33:33.120 20071 [Information] dongle is available: hci0
15:33:33.121 20071 [Error] I/O error: hciconfig hci0 up has failed: 72
15:33:33.121 20071 [Information] dongle is available: hci0
15:33:33.123 20071 [Error] I/O error: hciconfig hci0 up has failed: 72
15:33:33.123 20071 [Critical] dongle seems to be failing

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>