BelkinWemoDevice: new class
It is an abstract class through it can be accessed to belkin wemo
devices uniformly. The class defines that every device has DeviceID
and owns lock needed in case of dynamic allocation. Also determines
methods which have to be implemented by descended class. They are
methods for request modify state, request current state and get
properties of device.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>