HciInterface: reimplement up() by calling ioctl(HCIDEVUP)
Call ioctl(HCIDEVUP) to bring the selected interface UP. There is no need of the tool
hciconfig anymore. Anyway, the tool hciconfig is already deprecated and some systems
do not provide it.

Moreover, if the HCI interface is already UP, no root privileges are needed for this
command. This has impact on the BluetoothAvailabilityManager that can work without
root privileges if some other tool brings the HCI interface up before delivering the
particular hotplug event.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>