FitpDevice: new class
It keeps information (deviceID, type (end device, coordinator))
about devices that are paired within fitp network or are also
in the BeeeOn system.

If device exists, it is paired within fitp network. If device
has also set paired flag to true, it is paired in the BeeeOn system.

It also processes received data from coordinator or end device
and it is able to send module types of coordinator or end device.

Signed-off-by: Klara Necasova <xnecas24@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Klara Necasova <xnecas24@stud.fit.vutbr.cz>