Revert "base: update for BEEEON_OBJECT_TIME"
This reverts commit 17dc2fa849074564b93b7ccd3a58522878187dd4.

The base is preserved as is. The reason to revert this change is that
the base was originally moved in a wrong way to a non-merged commit.
Because the base has been moved in the meantime to master, the
TCPConsole modification is not available and he gateway fails to start
with testing.center.enable=yes.

This patch is to be reapplied after the base is updated correctly.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>