DataWriter: new class
Serves to write data, given by the DataIterator, with their checksum to the output stream.

Each data are written to the one line in the format:

  "HEXACHECKSUM""DATA"

  - "HEXACHECKSUM" is CRC-32 checksum of the following data in hexadecimal form
  - "DATA" is the exact data given by the DataIterator

This commis also includes DataWriterTest.

The test covers behaviour of the DataWriter in the regular case and
in the case of no data given.

Signed-off-by: Martin Hošala <xhosal00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>