GWServerConnector: implement pong frame receive
Implement receiving of the PONG frame. This frame is an expected answer
to the WebSocket PING frame. Sending of PING frames will be
introduced in upcoming patch.

Also add m_lastReceived timestamp to GWServerConnector, this Timestamp
represents when was the last message succesfully received. PONG frame
also updates this value.

receiveMessageUnlocked() can now return nullptr on PONG frame receive,
so the nullptr check needed to be implemented.

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>