UPnP: fix accepting of a device
It was added a check of the value of the ST header if it occurs in the received message.
If the value is the same as method's parametr deviceType accepts the device.
This is added due to Philips, because it responds to searching by messages containing the
ST header with all the values it knows. And it was also added to avoiding the return of
the same device multiple times.

ST header is more described in http://www.upnp.org/specs/arch/UPnP-arch-DeviceArchitecture-v1.0-20080424.pdf
on the page 19.

Fixes: de9a2c16 ("UPnP: new class")

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>