ZWaveCommandClassKey: fix operator <
The ZWaveCommandClassKey less-than comparator is wrong which leads
to wrong detection of supported data types via ZWaveClassRegistry.
IT was discovered that the support for battery sensor is always
considered as "unsupported" on real sensors.

This patch introduces tests of test ZWaveClassRegistry::operator <
and fixes this issue by first comparing class IDs, than if (and only
if) the IDs are same, the index is compared. Otherwise, the relation
does not apply.

Fixes: 68a6b345e120 ("ZWaveRegistry: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>