CommandDispatcher: replace virtual dispatch by dispatchImpl
The CommandDispatcher::dispatch() was declared virtual for inheritance.
However, in PocoCommandDispatcher::dispatch(), there is a prologue that
is generic enought to be put into all dispatch() implementations. Thus,
it is desirable to slightly change the way in which is a
CommandDispatcher implemented.

Now, the CommandDispatcher::dispatch() is not virtual and calls
internally the CommandDispatcher::dispatchImpl() method that is virtual.
All implementations thus override dispatchImpl().

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>