ZWaveDeviceManager: fix sometimes not reported unpaired devices
The pairing process is implemented in a really tricky way. There is
one important thing to consider. The notification Type_NodeQueriesComplete
is related to a single node. Also, if there is no _new_ node in the
Z-Wave network, this notification never comes. Thus, under such
circumstances, the dispatchUnpairedDevices() is not called and we
will not receive any information about unpaired devices.

This patch changes the discovery process behaviour:

1. Unpaired devices are dispatched upon Type_ControllerCommand
   with subtype ControllerState_Completed (and under error conditions).

2. The stopCommand() callback does not clear the m_state (only) in case
   of the discovery process (LISTENING) and leaves it to the notification
   handling state machine.

The step 2 is needed because we need to react only in the LISTENING mode.
Thus, the state machine now clears the m_state. Otherwise, the state machine
would respond also to other events (UNPAIRING) and do something unexpected.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>