Answer: convert to Poco::SynchronizedObject
The Poco::SynchronizedObject holds a recursive mutex inside. This
allows to simplify locking of objects like Answer where we sometimes
need to lock just a single operation, sometimes a sequence of operations
but we need to be always sure that we are locked properly.

As a result, we have to change the Result class because it reuses the
Answer lock and exposes it to anybody. For this reason, we define
methods lock() and unlock() for Result which are just redirected
to the appropriate Answer instance.

The change completely avoids the *Unlocked methods from Answer and
Result because they are not needed anymore.

Unfortunately, the Poco::SynchronizedObject and Poco::ScopedLock do
not support locking of const objects. Thus, const_cast must be used
time to time.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>