conf: provide syslog-ng configuration
There are 2 configuration files for syslog-ng:

1. udp-relay is intended to run next to the beeeon-gateway to forward
   log messages to some remote log collector.

2. tls-server configures a log collector as TLS server that stores
   logs into files based on the gateway ID stored in the gateway
   peer certificate.

There are some optional settings commented out. It is possible to
enable a more strict checking of clients/servers. The log collector
can define a whitelist of gateway IDs to accept.

The required-trusted option leads to set the SSL verify mode to

 SSL_VERIFY_PEER | SSL_VERIFY_CLIENT_ONCE | SSL_VERIFY_FAIL_IF_NO_PEER_CERT

For details on this topic, see SSL_set_verify(3). The "trusted"
also enables DN validation based on the trusted-dn setting.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>