BluetoothAvailabilityManager: detection of BLE devices
Detection extended by low energy scan.
For both detection and searching of BLE devices, a single method
HciInterface::lescan() is used. It returns a map of MACAddress and names
and each device is compared with list of paired devices.
Resolution between classic and LE device is solved in BluetoothDevice.

There is a cache of the last BLE scan result for fast result of listen
command. Cache can be old max LE_SCAN_TIME. In the case of dongle
problems, the cache is erased.

The BluetoothAvailabilityManager can be now configured in one of 3 modes:
* classic-only (bluetooth 3.0)
* low energy-only (bluetooth 4.0)
* both classic and low energy
* (none is also possible but not very interesting)

The mode is selected by calling method like setModes({"classic", "le"}).
For low energy is need a usb dongle with bluetooth 4.0 support.

Signed-off-by: Martin Sakin <xsakin00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>