VPTDevice: throw exception when the nonce is not found
When a VPT requires password to access it, the response should
contain a nonce which is required to create a password hash. If the
nonce is not found it means communication with a different device
than the VPT. In this case the exception is thrown for the better
error detection.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>