ZWavePocoLoggerAdapter: do not enforce log message format
The ZWavePocoLoggerAdapter is enforcing the message format by
prefixing all messages with "NodeId <ID>: " label. We can do
this better by setting a custom Poco::Message field to contain
the node ID. This way, the message contains more semantics and
the formatting is deferred to the appropriate FormattingChannel.
Thus, anybody can put the node ID anywhere into the log message.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>