DBusHciInterface: new class
This new class realizes communication with BlueZ deamon using D-Bus.
It allows to find new Bluetooth LE and Bluetooh Classic devices.

The methods for working with Bluetooth Classic devices uses the
BluezHciInterface class. The reason is that it is not possible to
detect Bluetooth Classic device using Bluez deamon when the device
is invisible.

Signed-off-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>