FitpDevice: fix TYPE_BATTERY to not use attributes
TYPE_BATTERY with ATTR_INNER does not make much sense. The INNER
and OUTER attributes describes measuring inside (a building) or
outside (in the wild). In case of batteries, we might consider
something like INTERNAL vs. EXTERNAL but there are no such attributes
defined and we do not need them now.

Fixes: 152a39f4a08d ("FitpDevice: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>