OZWCommand: new class
OZWCommand implements OpenZWave command management. The command
management is a bit tricky. In the current version, there is
only one way how to use it properly - to use the deprecated method
BeginControllerCommand. The method ensures that when updates
of the ControllerState are delivered, we *know* that they belong
to our command.

The same approach that uses the basic OpenZWave event loop (that
is implemented by OZWNetwork class) cannot be currently used
because the Notifications holding ControllerState do not contain
information about the command they belong to. This is however
changed in the development version of OpenZWave and should be
adapted later.

The current approach has probably some pitfalls. The ControllerStates
seem to not be delivered always as expected. Also, it is not clear
how exactly the OpenZWave library behaves when cancelling a command.

The methods running() and terminated() intentionally does not perform
any useful reactions on the current commands being in progress. This
is to be implemented in the following commit.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>