Answer: use own modification Event
Using the shared Event across all of the Answers
associated with the same AnswerQueue could lead to potential
race condition, causing dead-lock in the following situation.

When multiple threads dispatched a Command (i.e. N threads, N commands)
usign the same AnswerQueue, they all were blocked using same shared
Event. When any of those commands finished, the shared Event was notified
and a single one thread is woken up to check the status of
its Command. If a thread with an unfinished command was woken this way
then nothing happens (which might be wrong). Some other thread blocking
for a *finished* command might however not be woken up as expected. If no
other notification came, such thread would have hung, resulting in starvation
of the thread.

To resolve this problem, every Answer now has its own Event.
Also notification of this Event will notify shared Event in AnswerQueue.

Unit test was also updated accordingly, because it was depentant on shared Event,
which was notified directly, instead of using notifyUpdated().

Note that this problem could still occur when multiple threads are
waiting for the same instance of Answer to finish. Its not meant
to use like that anyway, and the warning note has been added to class
documentation.

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>