DeviceManager: implement method handle() generically
Implement generic method handle() that converts all exceptions
to FAILED result. All handling logic is to be implemented inside
the handleGeneric() method which is virtual and currently free
for overriding.

The method handleGeneric() might implement the commands in a generic
way or the particular device managers can override it for their own
specific purpose.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>