VirtualDeviceManager: FitpDeviceManager: fix handling unpair
When unpairing a non-existing device, it is not considered an error.
We should exit gracefully. This behaviour allows to run the unpair
chain of operations from server to gateway. If there is any issue
on the way (a bug, inconsistency, dead device, etc.), the unpair
command should succeed as much as possible.

If a device is already unpaired at the gateway level, it can still
be paired in other places (on the server). By failing here, we would
probably abort unpairing and disallow to remove such device from
other places (the server).

Fixes: cbe5f26cfb84 ("VirtualDeviceManager: new class")
Fixes: a46d2a427002 ("FitpDeviceManager: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>