PhilipsHueDeviceManager: migrate to handleGeneric()
We remove unnecessary try...catch from doListenCommand() which looks
as a big deal but in fact it is not.

Slightely more complex is modification of method doSetValueCommand().
We have to make sure that after successful modifyValue(). The ship()
must thus be protected there.

The method modifyValue() should be refactored to avoid returning bool
on behalf throwing exceptions.

Method doListenCommand() is racy because it does not lock.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>
Acked-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>