JablotronDeviceManager: migrate to generic handleListen()
The discovery process maintained by the JablotronDeviceManager is
implemented inside the main loop. Thus, the handleListen() that is
called from another thread just waits for new devices to be discovered.
Finally, when there are some, it simply dispatches them all at once.

We preserve this behaviour and adapt it to the AsyncWork interface.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>