VPTDeviceManager: migrate to generic handleListen()
The discovery process of the VPTDeviceManager is already asynchonous
via the VPTSeeker class. Thus, we just adapt the VPTSeeker to AsyncWork
interface.

We can share the VPTSeeker instance among calls to startDiscovery() as
it is guaranteed that the startDiscovery() is called and progressing
only once at a time and no other startDiscovery() is initiated until
the previous discovery finishes.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>