BluetoothAvailabilityManager: manage pairing state by DeviceCache
BluetoothAvailabilityManager uses quite a different approach among the implemented
device managers. We currently do not change the principle just make it closer to
the future approach that is based on the DeviceManager pairing management.

Pairing is controlled by methods addDevice(), removeDevice() and hasDevice(). Those
methods are used *only* while accepting, unpairing or discovering. The availability
detection *does not use those methods at all* and just iterates over the list of
added devices.

Thus, the hasDevice() method can be dropped (and we do so) as it can be easily
replaced by method DeviceCache::paired(). It has two possible (but inprobable)
side-effects:

1. If a device is marked as paired (by DeviceCache::markPaired()) and the
   m_deviceList is not updated, such device will not be reported during discovery
   and thus it might be impossible to pair it easily.

2. If a device is marked as unpaired (by DeviceCache::markUnpaired()) and the
   m_deviceList is not updated, such device would continue to ship data without
   any notice.

Both situations are not a big deal while we have no external access to the pairing
state enabled for the BluetoothAvailabilityManager. Otherwise, somebody has to
change that behaviour...

The BluetoothAvailabilityManager also forgets all paired devices on a failure
(removing dongle). This behaviour must be reworked in the future.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: David Bednařík <xbedna62@stud.fit.vutbr.cz>