Journal: new class
Journal generically implements the journaling principle for ensuring
fail-safe writing of data. A journal is a sequential structure of
records that is append-only. Time after time, it is possible to rebuild
the journal by dropping unneeded records from its beginning.

The class Journal works with records consisting of key and value.
The key refers to some entity and the value represents the entity's
value. The journal than maintains the entity's changes in time with
failure protection.

Example with entity "gateway.name":

  1) journal.append("gateway.name", "My Gateway");
  2) journal.append("gateway.name", "My best Gateway");
  3) journal.append("gateway.name", "My Cool Gateway");

If there is a (system, power) failure anywhere between 1) and 2) or
2) and 3), only the successfully written entry is considered during
recovery. If e.g. the 2) is written incompletely into the journal,
the journal would use 1).

Only the last record represents the current state of the entity.
Thus, if all records - 1), 2), and 3) - are written successfully,
then the "My Cool Gateway" is the only valid value of the key
"gateway.name".

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>